+++ ms2deepscore/train_new_model/__init__.py
@@ -1,10 +1,9 @@
-from .data_generators import DataGeneratorPytorch
-from .spectrum_pair_selection import (SelectedCompoundPairs,
-                                      select_compound_pairs_wrapper)
+from .data_generators import SpectrumPairGenerator, InchikeyPairGenerator
+from .inchikey_pair_selection import (select_compound_pairs_wrapper)
 
 
 __all__ = [
-    "DataGeneratorPytorch",
+    "SpectrumPairGenerator",
     "select_compound_pairs_wrapper",
-    "SelectedCompoundPairs",
+    "InchikeyPairGenerator"
 ]