+++ tests/test_validation_loss_calculator.py
@@ -1,12 +1,14 @@
+from collections import Counter
+
 import numpy as np
 import pytest
 from matchms.Spectrum import Spectrum
 from ms2deepscore.models.loss_functions import LOSS_FUNCTIONS
 from ms2deepscore.SettingsMS2Deepscore import SettingsMS2Deepscore
 from ms2deepscore.train_new_model.ValidationLossCalculator import (
-    ValidationLossCalculator, select_one_spectrum_per_inchikey)
+    ValidationLossCalculator, select_spectra_per_inchikey)
 from tests.create_test_spectra import (pesticides_test_spectra,
-                                       siamese_spectral_model)
+                                       siamese_spectral_model, create_test_spectra)
 
 
 @pytest.fixture()
@@ -35,23 +37,36 @@ def simple_test_spectra():
     return spectra
 
 
-def test_select_one_spectrum_per_inchikey(simple_test_spectra):
-    one_spectrum_per_inchikey = select_one_spectrum_per_inchikey(simple_test_spectra, 42)
-    assert len(one_spectrum_per_inchikey) == 2
+@pytest.mark.parametrize("nr_of_inchikeys,nr_of_spectra_per_inchikey,nr_of_sampled_spectra_per_inchikey",
+                         [[2, 2, 1],
+                          [2, 2, 5],
+                          [1, 2, 1],
+                          [2, 30, 100],])
+def test_select_one_spectrum_per_inchikey(nr_of_inchikeys, nr_of_spectra_per_inchikey,
+                                          nr_of_sampled_spectra_per_inchikey):
+    test_spectra = create_test_spectra(nr_of_inchikeys, nr_of_spectra_per_inchikey)
+    selected_spectra = select_spectra_per_inchikey(test_spectra, 42, nr_of_sampled_spectra_per_inchikey)
+    assert len(selected_spectra) == nr_of_inchikeys*nr_of_sampled_spectra_per_inchikey
 
     # Check if the spectra only are unique inchikeys
-    inchikeys_list = [s.get("inchikey") for s in one_spectrum_per_inchikey]
-    assert len(set(inchikeys_list)) == len(one_spectrum_per_inchikey), 'Expected 1 spectrum per inchikey. ' \
-                                                           'First run select_one_spectrum_per_inchikey'
-    # Check that the random seed works
-    assert one_spectrum_per_inchikey[0].get("precursor_mz") == 0.0
-    assert one_spectrum_per_inchikey[1].get("precursor_mz") == 17.0
+    inchikeys_list = [s.get("inchikey") for s in selected_spectra]
+    assert set(inchikeys_list) == set([s.get("inchikey") for s in test_spectra]), "not all inchikeys are selected"
+
+    for inchikey_count in Counter(inchikeys_list).values():
+        assert inchikey_count == nr_of_sampled_spectra_per_inchikey
+
+    hashed_spectra = [spectrum.set("fingerprint", None).__hash__() for spectrum in selected_spectra]
+    for spectrum_count in Counter(hashed_spectra).values():
+        minimum_spectrum_count = nr_of_sampled_spectra_per_inchikey // nr_of_spectra_per_inchikey
+        assert minimum_spectrum_count <= spectrum_count <= minimum_spectrum_count + 1, \
+            "The spectra are not sampled equally"
 
 
 def test_validation_loss_calculator():
     model = siamese_spectral_model()
     test_spectra = pesticides_test_spectra()
-    bins = np.array([(x / 10, x / 10 + 0.1) for x in range(0, 10)])
+    bins = np.array([(0.8, 0.9), (0.7, 0.8), (0.9, 1.0), (0.6, 0.7), (0.5, 0.6),
+                     (0.4, 0.5), (0.3, 0.4), (0.2, 0.3), (0.1, 0.2), (-0.01, 0.1)])
     validation_loss_calculator = ValidationLossCalculator(test_spectra,
                                                           settings=SettingsMS2Deepscore(same_prob_bins=bins))